Simple refactoring in packages and terminal menu formatting.
